% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randForest.R
\name{predict.randForest}
\alias{predict.randForest}
\title{predict.randForest: predict method for randForest objects}
\usage{
\method{predict}{randForest}(object, data, ...)
}
\arguments{
\item{object}{an object of class 'randForest', as that created by the function 'randForest'.}

\item{data}{Data frame containing new data.}
}
\value{
A vector of predicted values (0 or 1) is returned.
}
\description{
Prediction of test data using random forest.
}
\examples{
pred = predict(fit, df_test)
}
