% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randForest.R
\name{randForest}
\alias{randForest}
\title{Binary Classification with Random Forest}
\usage{
randForest(formula, data, ntree = 50, m_frac = 0.2)
}
\arguments{
\item{formula}{A data frame or a matrix of predictors, or a formula describing the model to be fitted.}

\item{data}{Data frame containing the variables in the model.}

\item{ntree}{Number of trees to grow. This should not be set to too small a number, to ensure that every input row gets predicted at least a few times.}

\item{m_frac}{Probability for each variable to be randomly sampled as candidate at each split.}
}
\value{
A list of ntree 'rpart' objects.
}
\description{
Build random forest for binary classification where all trees are created with the rpart package.
}
\examples{
fit = randForest(target ~ ., data=df_train, ntree=50, m_frac=0.1)
fit = randForest(satisfied ~ class + delay, data=planes, ntree=500, m_frac=0.5)
}
